# openapi.yaml
openapi: 3.0.3
info:
  title: POI Travel Recommendation API
  description: Flask API for POI management, route creation, and travel recommendations
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /api/search:
    get:
      summary: Advanced POI search
      operationId: searchPois
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/POI'
                  total:
                    type: integer
                  query:
                    type: string

  # Route endpoints
  /api/routes:
    get:
      summary: Get predefined routes
      operationId: getPredefinedRoutes
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Route'
                  total:
                    type: integer
                  page:
                    type: integer
                  total_pages:
                    type: integer
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    POI:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        images:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - latitude
        - longitude
...
    Route:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        estimated_duration:
          type: integer
          description: Duration in minutes
        difficulty:
          type: string
          enum: [easy, medium, hard]
        poi_count:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...